rag-knowledge-assistant/requirements.txt
langchain
openai
faiss-cpu
tiktoken
python-dotenv
langchain-community
pypdf
langchain_openai
sentence-transformers
langchain-ollama